Fixed empty lines and tests, deleted cleanup method
